# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order Management API",
        "description": "This API handles the management of customer orders",
        "contact": {
            "email": "aloshmohammad2001@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/catalogs/{catalogId}/colors": {
            "get": {
                "tags": [
                    "Colors"
                ],
                "summary": "Get colors of a catalog",
                "description": "Retrieves a paginated list of colors for a specific catalog by its ID",
                "operationId": "6bba1bb86fda5847786adb25137cb8ac",
                "parameters": [
                    {
                        "name": "catalogId",
                        "in": "path",
                        "description": "ID of the catalog for which colors are to be retrieved",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "Id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "Name": {
                                                                "type": "string",
                                                                "example": "Red"
                                                            },
                                                            "Quantity": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 100
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Colors"
                ],
                "summary": "Add colors to a catalog",
                "description": "Adds multiple colors to a specific catalog by its ID",
                "operationId": "9f9080f16287bf08419c3239a25300b6",
                "parameters": [
                    {
                        "name": "catalogId",
                        "in": "path",
                        "description": "ID of the catalog to which colors will be added",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "colors"
                                ],
                                "properties": {
                                    "colors": {
                                        "description": "List of colors to add",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "Name",
                                                "Quantity"
                                            ],
                                            "properties": {
                                                "Name": {
                                                    "type": "string",
                                                    "example": "red-test"
                                                },
                                                "Quantity": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 200
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Colors added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Color added successfully!"
                                        },
                                        "data": {
                                            "description": "List of added colors",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "CatalogId": {
                                                        "type": "integer",
                                                        "example": 18
                                                    },
                                                    "Name": {
                                                        "type": "string",
                                                        "example": "red-test"
                                                    },
                                                    "Quantity": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 200
                                                    },
                                                    "Id": {
                                                        "type": "integer",
                                                        "example": 26
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catalog not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Catalog not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/catalogs/{catalogId}/colors/update": {
            "post": {
                "tags": [
                    "Colors"
                ],
                "summary": "Update colors for a catalog",
                "description": "Updates the quantities of multiple colors for a specific catalog by its ID",
                "operationId": "d494439c825470d5feb45ca1d54e3487",
                "parameters": [
                    {
                        "name": "catalogId",
                        "in": "path",
                        "description": "ID of the catalog for which colors are to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "description": "List of colors with their quantities to update",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "Id",
                                                "Quantity"
                                            ],
                                            "properties": {
                                                "Id": {
                                                    "type": "integer",
                                                    "example": 26
                                                },
                                                "Quantity": {
                                                    "type": "integer",
                                                    "example": 20
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Colors updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Colors updated successfully"
                                        },
                                        "data": {
                                            "description": "List of updated colors",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "Id": {
                                                        "type": "integer",
                                                        "example": 26
                                                    },
                                                    "CatalogId": {
                                                        "type": "integer",
                                                        "example": 18
                                                    },
                                                    "Name": {
                                                        "type": "string",
                                                        "example": "red-test"
                                                    },
                                                    "Quantity": {
                                                        "type": "integer",
                                                        "example": 220
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catalog not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Catalog not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}